app/requirement.txt
fastapi
pydantic
python-dotenv
sqlalchemy
aiomysql
motor
langgraph
langchain-core
langchain-groq
typing-extensions
uvicorn